backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kasper Junge", email = "kasperjuunge@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "fastapi-users[oauth,sqlalchemy]>=14.0.1",
    "asyncpg>=0.30.0",
    "fastapi-users-db-sqlalchemy>=7.0.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.11",
]
